% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDNS2010.R
\docType{data}
\name{IDNS2010}
\alias{IDNS2010}
\title{IDNS2010}
\format{a data frame with five variables
\describe{
\item{ID}{ID, in numeric}
\item{Provinces}{Provinces in Indonesia}
\item{PopulationIDN}{Population in Indonesia, in numeric}
\item{TotalHousehold}{TotalHousehold in Indonesia, in numeric}
\item{PopulationMen}{PopulationMen, in numeric}
\item{PopulationwoMen}{PopulationwoMen, in numeric}
\item{Household}{Household, in numeric}
\item{Incomeff}{Incomeff, in numeric}
\item{Incomemerchant}{Incomemerchant, in numeric}
\item{Incomegov}{Incomegov, in numeric}
\item{Incomepriv}{Incomepriv, in numeric}
\item{Incomeentrep}{Incomeentrep, in numeric}
\item{Incomefreelancer}{Incomefreelancer, in numeric}
\item{Medianage}{Medianage, in numeric}
\item{RatioParticp}{RatioParticp, in numeric}
\item{Unemployment2015}{Unemployment2015, in numeric}
\item{Unemployment2016}{Unemployment2016, in numeric}
\item{Unemployment2017}{Unemployment2017, in numeric}
\item{PopulationProjection2010}{PopulationProjection2010, in numeric}
\item{PopulationProjection2015}{PopulationProjection2015, in numeric}
\item{PopulationProjection2020}{PopulationProjection2020, in numeric}
\item{PopulationProjection2025}{PopulationProjection2025, in numeric}
\item{PopulationProjection2030}{PopulationProjection2030, in numeric}
\item{PopulationProjection2035}{PopulationProjection2035, in numeric}
\item{No}{no, in numeric}

}}
\source{
\url{https://www.bps.go.id/}

  downloaded from the source url on 28 May 2019
}
\usage{
IDNS2010
}
\description{
These are the census data about, data from statistics office that used to see the correlation of the census and election data
}
\keyword{datasets}
