% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Legislative_Election.R
\docType{data}
\name{Legislative_Election}
\alias{Legislative_Election}
\title{Legislative_Election}
\format{a data frame with three variables
\describe{
\item{Names}{Names}
\item{PartyNm}{Party}
\item{Electorate}{Electorate in Indonesia}}}
\source{
\url{http://www.tribunnews.com/pemilu-2014/2014/05/15/daftar-caleg-yang-lolos-ke-dpr-ri}
  downloaded from the source url on 28 May 2019
}
\usage{
Legislative_Election
}
\description{
Datasets used to calculated the party that won the legislative election, there are three variables which used to calculated the party that won.
}
\keyword{datasets}
