% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalvotes.R
\docType{data}
\name{totalvotes}
\alias{totalvotes}
\title{totalvotes}
\format{a data frame with five variables
\describe{
\item{Names}{Names}
\item{PartyNm}{Party}
\item{Electorate}{Electorate in Indonesia}
\item{Provinces}{provinces in Indonesia}
\item{Votes}{Total votes, in numeric}
}}
\source{
\url{http://www.tribunnews.com/pemilu-2014/2014/05/15/daftar-caleg-yang-lolos-ke-dpr-ri}

  downloaded from the source url on 28 May 2019
}
\usage{
totalvotes
}
\description{
Datasets used to calculated the party receiving the most votes in the legislative election, which candidate did the best and how did each electorate vote in each province.
}
\keyword{datasets}
